Fix: Ajuste no refactor definindo um package para cada funcionalidade no código (#38)
