+++ src/main/java/net/weg/repository/TesteConexaoBanco.java
@@ -1,4 +1,4 @@
-package org.example;
+package net.weg.repository;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
