+++ src/main/java/net/weg/services/ManutencaoReparo.java
@@ -1,4 +1,4 @@
-package org.example;
+package net.weg.services;
 
 import java.util.Scanner;
