+++ src/main/java/net/weg/services/ProcedimentodeSeguranca.java
@@ -1,12 +1,12 @@
-package org.example;
+package net.weg.services;
 
 import java.util.Scanner;
 
-public class TesteDiagnostico {
+public class ProcedimentodeSeguranca {
 
     String nomeProduto;
+    String comoUsar;
     String exemploPratico;
-    String descricaoDiagnostico;
 
     public void setNomeProduto(String nomeProduto) {
 
@@ -14,49 +14,47 @@ public class TesteDiagnostico {
 
     }
 
-    public void setexemploPratico(String exemploPratico) {
+    public void setComoUsar(String comoUsar) {
+
+        this.comoUsar = comoUsar;
+
+    }
+
+    public void setExemploPratico(String exemploPratico) {
 
         this.exemploPratico = exemploPratico;
 
     }
 
-    public void setDescricaoDiagnostico(String descricaoDiagnostico) {
-
-        this.descricaoDiagnostico = descricaoDiagnostico;
-
-    }
-
     public String toString() {
 
-        return "\n==================================================================================================================================="
-                + "\nNome Orientação: " + this.nomeProduto
+        return  "================================================================================================================================="
+                + "Nome Orientação: " + this.nomeProduto
+                + "\n----------------------------------------------------------------------------------------------------------------------------------"
+                + "\nComo Usar: " + this.comoUsar
                 + "\n----------------------------------------------------------------------------------------------------------------------------------"
                 + "\nExemplo Pratico: " + this.exemploPratico
-                + "\n----------------------------------------------------------------------------------------------------------------------------------"
-                + "\nDescrição diagnostico: " + this.descricaoDiagnostico
                 + "\n=================================================================================================================================";
     }
 
-    public String cadastrarProdutoDiagnostico() {
+    public String cadastrarProdutoSegurança() {
         Scanner scanner = new Scanner(System.in);
 
         System.out.print("\nDigite o nome do produto: ");
         this.nomeProduto = scanner.nextLine();
 
         System.out.print("Digite a descrição do produto: ");
-        this.descricaoDiagnostico = scanner.nextLine();
+        this.comoUsar = scanner.nextLine();
 
         System.out.print("Digite um exemplo prático: ");
         this.exemploPratico = scanner.nextLine();
 
         System.out.println("\nOrientação cadastrada com sucesso!");
 
-
-        return "Produto cadastrado: " + this.nomeProduto + " - " + this.descricaoDiagnostico + "-" + this.exemploPratico;
+        return "Produto cadastrado: " + this.nomeProduto + " - " + this.comoUsar + "-" + this.exemploPratico;
 
 
     }
 
 
 }
-
