+++ src/main/java/net/weg/services/ProjectJava.java
@@ -1,4 +1,4 @@
-package org.example;
+package net.weg.services;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
